// types/supabase.ts
export type Database = {
// ...
  public: {
    Tables: {
      credits: {
        Row: {
          created_at: string
          credits: number
          id: number
          user_id: string
        }
        Insert: {
          created_at?: string
          credits?: number
          id?: number
          user_id: string
        }
        Update: {
          created_at?: string
          credits?: number
          id?: number
          user_id?: string
        }
        Relationships: []
      }
      generation_jobs: {
        Row: {
          created_at: string
          id: number
          modelId: string | null
          name: string | null
          status: string
          type: string | null
          user_id: string | null
        }
        Insert: {
          created_at?: string
          id?: number
          modelId?: string | null
          name?: string | null
          status?: string
          type?: string | null
          user_id?: string | null
        }
        Update: {
          created_at?: string
          id?: number
          modelId?: string | null
          name?: string | null
          status?: string
          type?: string | null
          user_id?: string | null
        }
        Relationships: []
      }
      images: {
        Row: {
          background_preset: string | null
          created_at: string
          id: number
          is_favorited: boolean | null
          modelId: number
          parent_image_id: number | null
          source: string | null
          style_preset: string | null
          uri: string
        }
        Insert: {
          background_preset?: string | null
          created_at?: string
          id?: number
          is_favorited?: boolean | null
          modelId: number
          parent_image_id?: number | null
          source?: string | null
          style_preset?: string | null
          uri: string
        }
        Update: {
          background_preset?: string | null
          created_at?: string
          id?: number
          is_favorited?: boolean | null
          modelId?: number
          parent_image_id?: number | null
          source?: string | null
          style_preset?: string | null
          uri?: string
        }
        Relationships: [
          {
            foreignKeyName: "images_modelId_fkey"
            columns: ["modelId"]
            isOneToOne: false
            referencedRelation: "generation_jobs"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "images_parent_image_id_fkey"
            columns: ["parent_image_id"]
            isOneToOne: false
            referencedRelation: "images"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
// ...
}